requirements.txt
about-time==4.2.1
alive-progress==3.1.4
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.3
grapheme==0.6.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imutils==0.5.4
moviepy==2.0.0.dev2
natsort==8.4.0
numpy==1.25.0
opencv-python==4.7.0.72
pandas==2.0.3
Pillow==10.0.0
proglog==0.1.10
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
